Moving update_configs into DqConfigsCache
