+++ clouddq/classes/dq_entity.py
@@ -104,10 +104,6 @@ class DqEntity:
     dataplex_createTime: str | None
     dataplex_updateTime: str | None
 
-    @classmethod
-    def update_config(cls: DqEntity, config_current: dict, config_new: dict) -> dict:
-        return clouddq.classes.update_config(config_current, config_new)
-
     def resolve_column_config(self: DqEntity, column_id: str) -> DqEntityColumn:
         """
