+++ clouddq/lib.py
@@ -49,8 +49,8 @@ def load_configs(configs_path: Path, configs_type: DqConfigType) -> typing.Dict:
         if not config:
             continue
 
-        all_configs = DqConfigType.to_class(configs_type).update_config(
-            all_configs, config
+        all_configs = DqConfigsCache.update_config(
+            configs_type, all_configs, config
         )
 
     if configs_type.is_required():
